+++ apps/driva_editor/test/modules/editor_module/presentation/editor/editor_perf_test.dart
@@ -1,4 +1,5 @@
 import 'package:bloc_test/bloc_test.dart';
+import 'package:driva_editor/core/theme/app_theme.dart';
 import 'package:driva_editor/modules/editor_module/domain/use_cases/use_cases.dart';
 import 'package:driva_editor/modules/editor_module/presentation/editor/cubit/editor_cubit.dart';
 import 'package:driva_editor/modules/editor_module/presentation/editor/editor_page.dart';
@@ -50,6 +51,7 @@ void main() {
   tearDown(() => cubit.close());
 
   Widget harness() => MaterialApp(
+    theme: AppTheme.light,
     home: MultiBlocProvider(
       providers: [
         BlocProvider<EditorCubit>.value(value: cubit),